Updated compatibility
I think the API changed at Internet Archive in some way that broke Charles' old (slicker) way of dynamically getting video lists.  I punted and hard coded it, which saves a little traffic to IA and a little startup time.  Smarter shuffling so no duplicate videos on the screen at the same time and reduced the number of simultaneou videos to also take it a little easier on IA.